refactor: getImageUrl 함수를 httpAction으로 변경하고 http 라우터에 통합
