+++ convex/files.ts
@@ -1,7 +1,8 @@
 import { zid } from 'convex-helpers/server/zod';
-import { zMutation, zQuery } from '~/shared/lib/zod-convex';
+import { zMutation } from '~/shared/lib/zod-convex';
 
-import { mutation } from './_generated/server';
+import type { Id } from './_generated/dataModel';
+import { httpAction, mutation } from './_generated/server';
 
 export const generateUploadUrl = mutation(async (ctx) => {
   return await ctx.storage.generateUploadUrl();
@@ -17,9 +18,15 @@ export const sendImage = zMutation({
   },
 });
 
-export const getImageUrl = zQuery({
-  args: { storageId: zid('_storage') },
-  handler: async (ctx, args) => {
-    return await ctx.storage.getUrl(args.storageId);
-  },
+export const getImageUrl = httpAction(async (ctx, request) => {
+  const { searchParams } = new URL(request.url);
+  const storageId = searchParams.get('storageId')! as Id<'_storage'>;
+  const blob = await ctx.storage.get(storageId);
+
+  if (blob === null) {
+    return new Response('Image not found', {
+      status: 404,
+    });
+  }
+  return new Response(blob);
 });
